% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver_add_to_rprofile.R
\name{languageserver_add_to_rprofile}
\alias{languageserver_add_to_rprofile}
\title{Add languageserver initialization to Rprofile}
\usage{
languageserver_add_to_rprofile(rlsLib = getOption("langserver_library"),
  rprofilePath = locate_rprofile(), confirmBeforeChanging = TRUE,
  code = append_code(rlsLib = rlsLib))
}
\arguments{
\item{rlsLib}{\code{character(1)} path to the library.}

\item{rprofilePath}{\code{character(1)}, path to the file where
to add the initialization code, or \code{NULL}. By default, adds the
code to a \code{.Rprofile} file in the home directory of the current
user.}

\item{confirmBeforeChanging}{\code{logical(1)}, if \code{TRUE}, aks for user
confirmation before changing the file. For non-interactive
use, \code{FALSE} will skip the confirmation.}

\item{code}{\code{character()}, the code to be added to the file.
Defaults to the value of \code{append_code()}.}
}
\value{
side-effects
}
\description{
Add languageserver initialization to Rprofile
}
