% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{languageserver_install}
\alias{languageserver_install}
\title{Install language server to a separate library}
\usage{
languageserver_install(rlsLib = getOption("langserver_library"),
  strictLibrary = TRUE, fullReinstall = TRUE, fromGitHub = TRUE,
  ref = "master")
}
\arguments{
\item{rlsLib}{\code{character(1)} path to the library.}

\item{strictLibrary}{\code{logical(1)} if \code{TRUE}, all the dependencies
of \code{languageserver} will be installed into \code{rlsLib}, otherwise
only those that are needed but not present in other libraries
in \code{.libPaths()} will be installed.}

\item{fullReinstall}{\code{logical(1)}. If \code{TRUE}, \code{rlsLib} will be
recursively removed to re-install all the packages cleanly.}

\item{fromGitHub}{\code{logical(1)}, if \code{TRUE}, will use
\code{remotes::install_github()}, otherwise \code{install.packages()} is
used to install the \code{languageserver} package.}

\item{ref}{\code{character(1)}, passed to \code{remotes::install_github()}
when relevant.}
}
\value{
side-effects
}
\description{
Install language server to a separate library
}
