% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver_remove_from_rprofile.R
\name{languageserver_remove_from_rprofile}
\alias{languageserver_remove_from_rprofile}
\title{Remove languageserver initialization from Rprofile}
\usage{
languageserver_remove_from_rprofile(rprofilePath = locate_rprofile(),
  code = append_code(), confirmBeforeChanging = TRUE)
}
\arguments{
\item{rprofilePath}{\code{character(1)}, path to the file where
to add the initialization code, or \code{NULL}. By default, adds the
code to a \code{.Rprofile} file in the home directory of the current
user.}

\item{code}{\code{character()}, the code to be added to the file.
Defaults to the value of \code{append_code()}.}

\item{confirmBeforeChanging}{\code{logical(1)}, if \code{TRUE}, aks for user
confirmation before changing the file. For non-interactive
use, \code{FALSE} will skip the confirmation.}
}
\value{
side-effects
}
\description{
Remove languageserver initialization from Rprofile
}
