% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{languageserver_startup}
\alias{languageserver_startup}
\title{Prepare languageserver to be started}
\usage{
languageserver_startup(rlsLib = getOption("langserver_library"),
  langServerProcessPatt = getOption("langserver_processPatt"),
  strictLibrary = TRUE, processArgs = list(command = "ps", args = c("-p",
  Sys.getpid(), "-o", "cmd", "--no-headers"), stdout = TRUE))
}
\arguments{
\item{rlsLib}{\code{character(1)} path to the library.}

\item{langServerProcessPatt}{\code{character(1)}, pattern to
recognize the process created by language server.}

\item{strictLibrary}{\code{logical(1)} if \code{TRUE}, all the dependencies
of \code{languageserver} will be installed into \code{rlsLib}, otherwise
only those that are needed but not present in other libraries
in \code{.libPaths()} will be installed.}

\item{processArgs}{\code{list()} of arguments to \code{system2} to
retrieve a list of processes.}
}
\description{
Prepare languageserver to be started
}
